Conflicts:
	data/xl.mdf
	data/xl_log.ldf

	new file:   libs/server2003/msado15.dll
	modified:   src/TransServer/MsSqlServer.cpp
	modified:   src/TransServer/MsSqlServer.h
	modified:   src/TransServer/XlCType.h
	modified:   src/TransServer/XlClient.cpp
	modified:   src/TransServer/XlDType.h
	modified:   src/TransServer/XlHelper.cpp
	modified:   src/TransServer/XlHost.cpp
	modified:   src/TransServer/XlHost.h
	modified:   src/client_demo/ClientImpl.cpp
	modified:   src/client_demo/VideoDlg.cpp
	modified:   src/client_demo/XlClientType.h

+++ src/client_demo/XlClientType.h
@@ -231,7 +231,7 @@ typedef struct _alarmInfoReq
 typedef struct _alarmInfoResp
 {
 	char hostId[32];								///< 设备信息
-	__int64 bAlarm;								///< 报警信息
+	int bAlarm;								///< 报警信息
 	struct  
 	{
 		double dLatitude;
